% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_X_ed_hist}
\alias{create_X_ed_hist}
\title{Creates ed_hist database table with id and first_noted columns}
\usage{
create_X_ed_hist(conn, X_ed_hist)
}
\arguments{
\item{conn}{database connection object}

\item{X_ed_hist}{character vector of length one with the name of the table to
create}
}
\description{
Creates ed_hist database table with id and first_noted columns
}

