% Generated by roxygen2 (4.0.1): do not edit by hand
\name{define_corral_locations}
\alias{define_corral_locations}
\title{Returns a numeric vector with the locations that are classified as corral
housing structures.}
\usage{
define_corral_locations(conn)
}
\arguments{
\item{conn}{database connection object}
}
\description{
Returns a numeric vector with the locations that are classified as corral
housing structures.
}

