% Generated by roxygen2 (4.0.1): do not edit by hand
\name{define_gang_locations}
\alias{define_gang_locations}
\title{Returns a numeric vector with the locations that are classified as gang
housing structures.}
\usage{
define_gang_locations(conn)
}
\arguments{
\item{conn}{database connection object}
}
\description{
Returns a numeric vector with the locations that are classified as gang
housing structures.
}

