% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_days_gang}
\alias{get_days_gang}
\title{Returns the number of days a specific animal was in gang cages.}
\usage{
get_days_gang(id, daily_df)
}
\arguments{
\item{id}{animal id}

\item{daily_df}{dataframe containing the animal and location data}
}
\description{
Returns the number of days a specific animal was in gang cages.
}

