% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_days_single}
\alias{get_days_single}
\title{Returns the number of days a specific animal was in a single housing
location.}
\usage{
get_days_single(id, new_df)
}
\arguments{
\item{id}{animal id}

\item{new_df}{dataframe containing the animal and location data}
}
\description{
Returns the number of days a specific animal was in a single housing
location.
}

