% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_housing_types}
\alias{get_housing_types}
\title{Returns a list of length 3 with a vector of gang cage locations, a vector
of corral locations, and a vector of single locations.}
\usage{
get_housing_types(conn)
}
\arguments{
\item{conn}{database connection object}
}
\description{
Returns a list of length 3 with a vector of gang cage locations, a vector
of corral locations, and a vector of single locations.
}

