% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_male_female_ratio}
\alias{get_male_female_ratio}
\title{Returns a dataframe with the Id, date, and the number of males and females
of a specified species on that date.}
\usage{
get_male_female_ratio(conn, affected_df, arc_species_code)
}
\arguments{
\item{conn}{database connection object}

\item{affected_df}{dataframe containing data about affected animals (e.g.,
id, sex, date Histoplasmosis was first noted, etc.)}

\item{target_date_df}{dataframe with the dates (target_date)}

\item{arc_species_code}{of all animals being counted.}
}
\description{
Returns a dataframe with the Id, date, and the number of males and females
of a specified species on that date.
}

