% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_new_df}
\alias{make_new_df}
\title{Creates a dataframe with a row for each day of life prior to the date
Histoplasmosis was first noted for each animal in the table with the name
contained in X_ed_hist.}
\usage{
make_new_df(conn, X_ed_hist)
}
\arguments{
\item{conn}{database connection object}

\item{X_ed_hist}{name of database table with Ids and dates Histoplasmosis
was first noted.}
}
\description{
Creates a dataframe with a row for each day of life prior to the date
Histoplasmosis was first noted for each animal in the table with the name
contained in X_ed_hist.
}

