% Generated by roxygen2 (4.0.1): do not edit by hand
\name{insert_id_first_noted}
\alias{insert_id_first_noted}
\title{Inserts ids and dates that correspond to when the Histoplasmosis infection
was first noted.}
\usage{
insert_id_first_noted(conn, X_ed_hist, df)
}
\arguments{
\item{conn}{database connection object}

\item{X_ed_hist}{name of database table receiving inserts}

\item{df}{dataframe containing data being inserted.}
}
\description{
Inserts ids and dates that correspond to when the Histoplasmosis infection
was first noted.
}

